Bugfix and cleanup for the validation feedback display

+++ src/utils.js
@@ -89,13 +89,15 @@ export function argsDictFromObject(args) {
  * @param {string} dir - path to a directory
  */
 export function cleanupDir(dir) {
-  fs.readdirSync(dir).forEach((filename) => {
-    const filepath = path.join(dir, filename);
-    if (fs.lstatSync(filepath).isFile()) {
-      fs.unlinkSync(filepath);
-    } else {
-      cleanupDir(filepath);
-    }
-  });
-  fs.rmdirSync(dir);
+  if (fs.existsSync(dir)) {
+    fs.readdirSync(dir).forEach((filename) => {
+      const filepath = path.join(dir, filename);
+      if (fs.lstatSync(filepath).isFile()) {
+        fs.unlinkSync(filepath);
+      } else {
+        cleanupDir(filepath);
+      }
+    });
+    fs.rmdirSync(dir);
+  }
 }
